backend/requirements.txt
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==2.6.0
six==1.16.0
SQLAlchemy==1.4.46
uuid==1.30
Werkzeug==2.2.2
